Plugins: move WpcomPluginInstallButton out of plugins-wpcom to its own directory
The `plugins-wpcom` directory will be completely deleted soon, so let's move the few
needed components from there to another location.
