+++ client/my-sites/plugins/plugin-install-button-wpcom/index.jsx
